requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
Django==3.2.10
django-appconf==1.0.5
django-autocomplete-light==3.8.2
django-cors-headers==3.10.1
django-cryptography==1.0
django-random-queryset==0.1.3
django-registration-redux==2.9
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
idna==3.3
pathlib==1.0.1
Pillow==8.4.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyotp==2.6.0
python-decouple==3.5
python-magic==0.4.24
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
